+++ Project/PayRoad/DateUtil.swift
@@ -39,21 +39,3 @@ struct DateUtil {
         return ymds
     }
 }
-
-extension DateFormatter {
-    static func stringToDay(for date: Date) -> String {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "dd"
-        
-        return dateFormatter.string(from: date)
-    }
-    
-    static func string(for date: Date, timeZone: TimeZone) -> String {
-        let formatter = DateFormatter()
-        formatter.timeZone = timeZone
-        formatter.dateStyle = .medium
-        formatter.timeStyle = .medium
-        
-        return formatter.string(from: date)
-    }
-}
